Move autosave components to their own files
