+++ spyder/plugins/editor/tests/test_autosave.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+#
+"""Tests for autosave.py"""
+
+# Third party imports
+import pytest
+
+# Local imports
+from spyder.plugins.editor.autosave import AutosaveComponentForEditorPlugin
+
+
+def test_autosave_component_set_interval(qtbot, mocker):
+    """Test that setting the interval does indeed change it and calls
+    do_autosave if enabled."""
+    mocker.patch.object(AutosaveComponentForEditorPlugin, 'do_autosave')
+    addon = AutosaveComponentForEditorPlugin(None)
+    addon.do_autosave.assert_not_called()
+    addon.interval = 10000
+    assert addon.interval == 10000
+    addon.do_autosave.assert_not_called()
+    addon.enabled = True
+    addon.interval = 20000
+    assert addon.do_autosave.called
+
+
+@pytest.mark.parametrize('enabled', [False, True])
+def test_autosave_component_timer_if_enabled(qtbot, mocker, enabled):
+    """Test that AutosaveCompenent calls do_autosave() on timer if enabled."""
+    mocker.patch.object(AutosaveComponentForEditorPlugin, 'do_autosave')
+    addon = AutosaveComponentForEditorPlugin(None)
+    addon.do_autosave.assert_not_called()
+    addon.interval = 100
+    addon.enabled = enabled
+    qtbot.wait(500)
+    if enabled:
+        assert addon.do_autosave.called
+    else:
+        addon.do_autosave.assert_not_called()
